+++ src/python/pants/backend/core/BUILD
@@ -6,7 +6,7 @@ python_library(
   name = 'plugin',
   sources = ['register.py'],
   dependencies = [
-    'src/python/pants/backend/core/targets:all',
+    'src/python/pants/backend/core/targets:doc',
     'src/python/pants/backend/core/tasks:all',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:source_root',
@@ -37,7 +37,6 @@ python_library(
   ],
   dependencies = [
     '3rdparty/python:six',
-    'src/python/pants/backend/core/targets:all',
     'src/python/pants/backend/core/tasks:all',
     'src/python/pants/base:build_environment',
     'src/python/pants/build_graph',
