+++ src/python/pants/backend/core/targets/resources.py
@@ -5,36 +5,16 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-from pants.base.payload import Payload
-from pants.build_graph.target import Target
+from pants.base.deprecated import deprecated_module
+from pants.build_graph.resources import Resources
 
 
-class Resources(Target):
-  """JVM resource files.
+deprecated_module('0.0.66',
+                  hint_message='pants.backend.core.targets.resources has moved to '
+                               'pants.build_graph.resources. Replace deps on '
+                               'src/python/pants/backend/core/targets:all or '
+                               'src/python/pants/backend/core/targets:common with a dep on '
+                               'src/python/pants/build_graph and change imports accordingly.')
 
-  Looking for loose files in your application bundle? Those are
-  `bundle <#bundle>`_\s.
 
-  Resources are Java-style resources accessible via the ``Class.getResource``
-  and friends API. In the ``jar`` goal, the resource files are placed in the resulting `.jar`.
-  """
-
-  def __init__(self, address=None, payload=None, sources=None, **kwargs):
-    """
-    :param sources: Files to "include". Paths are relative to the
-      BUILD file's directory.
-    :type sources: ``Fileset`` or list of strings
-    """
-    payload = payload or Payload()
-    payload.add_fields({
-      'sources': self.create_sources_field(sources,
-                                           sources_rel_path=address.spec_path, key_arg='sources'),
-    })
-    super(Resources, self).__init__(address=address, payload=payload, **kwargs)
-
-  def has_sources(self, extension=None):
-    """``Resources`` never own sources of any particular native type, like for example
-    ``JavaLibrary``.
-    """
-    # TODO(John Sirois): track down the reason for this hack and kill or explain better.
-    return extension is None
+Resources = Resources
