+++ src/python/pants/backend/core/tasks/BUILD
@@ -67,9 +67,10 @@ python_library(
   sources = ['confluence_publish.py'],
   dependencies = [
     '3rdparty/python/twitter/commons:twitter.common.confluence',
-    'src/python/pants/backend/core/targets:common',
+    'src/python/pants/backend/core/targets:doc',
     'src/python/pants/base:exceptions',
     'src/python/pants/binaries:binary_util',
+    'src/python/pants/build_graph',
     'src/python/pants/task',
     'src/python/pants/util:dirutil',
   ],
@@ -127,7 +128,7 @@ python_library(
     '3rdparty/python:Pygments',
     '3rdparty/python:docutils',
     '3rdparty/python:six',
-    'src/python/pants/backend/core/targets:common',
+    'src/python/pants/backend/core/targets:doc',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:generator',
