+++ src/python/pants/backend/jvm/targets/jvm_target.py
@@ -7,7 +7,6 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 from twitter.common.collections import OrderedSet
 
-from pants.backend.core.targets.resources import Resources
 from pants.backend.jvm.subsystems.jvm_platform import JvmPlatform
 from pants.backend.jvm.targets.exclude import Exclude
 from pants.backend.jvm.targets.jar_library import JarLibrary
@@ -15,6 +14,7 @@ from pants.backend.jvm.targets.jarable import Jarable
 from pants.base.exceptions import TargetDefinitionException
 from pants.base.payload import Payload
 from pants.base.payload_field import ExcludesField, PrimitiveField
+from pants.build_graph.resources import Resources
 from pants.build_graph.target import Target
 from pants.util.memo import memoized_property
