+++ src/python/pants/backend/jvm/tasks/jvm_dependency_usage.py
@@ -10,11 +10,11 @@ import os
 import sys
 from collections import defaultdict, namedtuple
 
-from pants.backend.core.targets.resources import Resources
 from pants.backend.jvm.targets.jar_library import JarLibrary
 from pants.backend.jvm.tasks.classpath_util import ClasspathUtil
 from pants.backend.jvm.tasks.jvm_dependency_analyzer import JvmDependencyAnalyzer
 from pants.base.build_environment import get_buildroot
+from pants.build_graph.resources import Resources
 from pants.build_graph.target import Target
 from pants.util.dirutil import fast_relpath
 from pants.util.fileutil import create_size_estimators
