+++ src/python/pants/backend/project_info/tasks/BUILD
@@ -83,7 +83,6 @@ python_library(
   dependencies = [
     '3rdparty/python/twitter/commons:twitter.common.collections',
     '3rdparty/python:pex',
-    'src/python/pants/backend/core/targets:common',
     'src/python/pants/backend/jvm/subsystems:jvm_platform',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm/targets:scala',
@@ -94,6 +93,7 @@ python_library(
     'src/python/pants/backend/python/tasks:python',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
+    'src/python/pants/build_graph',
     'src/python/pants/java/distribution',
     'src/python/pants/java:executor',
     'src/python/pants/option',
@@ -108,8 +108,8 @@ python_library(
   dependencies = [
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm/targets:scala',
-    'src/python/pants/backend/core/targets:common',
     'src/python/pants/base:build_environment',
+    'src/python/pants/build_graph',
     'src/python/pants/task',
   ],
 )
@@ -119,7 +119,6 @@ python_library(
   sources = ['ide_gen.py'],
   dependencies = [
     '3rdparty/python/twitter/commons:twitter.common.collections',
-    'src/python/pants/backend/core/targets:common',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:scala',
     'src/python/pants/backend/jvm/tasks:classpath_products',
