+++ src/python/pants/backend/python/targets/BUILD
@@ -15,7 +15,6 @@ python_library(
     '3rdparty/python:setuptools',
     '3rdparty/python:six',
     '3rdparty/python/twitter/commons:twitter.common.collections',
-    'src/python/pants/backend/core/targets:common',
     'src/python/pants/backend/python:python_artifact',
     'src/python/pants/backend/python:python_requirement',
     'src/python/pants/base:exceptions',
