+++ src/python/pants/docs/howto_develop.md
@@ -190,13 +190,13 @@ use the usual `./pants py` on a `python_library` target that builds (or
 depends on) the modules you want:
 
     :::bash
-    $ ./pants py src/python/pants/backend/core/targets:common
-    /Users/lhosken/workspace/pants src/python/pants/backend/core/targets:common
+    $ ./pants py src/python/pants/build_graph
+    /Users/lhosken/workspace/pants src/python/pants/build_graph
     Python 2.6.8 (unknown, Mar  9 2014, 22:16:00)
     [GCC 4.2.1 Compatible Apple LLVM 5.0 (clang-500.0.68)] on darwin
     Type "help", "copyright", "credits" or "license" for more information.
     (InteractiveConsole)
-    >>> from pants.backend.core.targets import repository
+    >>> from pants.build_graph.target import Target
     >>>
 
 Developing and Debugging a JVM Tool
