+++ tests/python/pants_test/backend/core/tasks/BUILD
@@ -40,7 +40,6 @@ python_tests(
   sources = ['test_dependees.py'],
   dependencies = [
     'src/python/pants/backend/codegen/targets:java',
-    'src/python/pants/backend/core/targets:common',
     'src/python/pants/backend/core/tasks:dependees',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:jvm',
@@ -54,7 +53,7 @@ python_tests(
   name = 'filter',
   sources = ['test_filter.py'],
   dependencies = [
-    'src/python/pants/backend/core/targets:common',
+    'src/python/pants/backend/core/targets:doc',
     'src/python/pants/backend/core/tasks:filter',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/python/targets:python',
@@ -70,9 +69,10 @@ python_tests(
   dependencies = [
     '3rdparty/python:beautifulsoup4',
     '3rdparty/python:mock',
-    'src/python/pants/backend/core/targets:common',
+    'src/python/pants/backend/core/targets:doc',
     'src/python/pants/backend/core/tasks:markdown_to_html',
     'src/python/pants/base:exceptions',
+    'src/python/pants/build_graph',
     'tests/python/pants_test/tasks:task_test_base',
   ],
   coverage=[
