+++ tests/python/pants_test/backend/jvm/tasks/test_jar_create.py
@@ -10,12 +10,12 @@ from contextlib import closing
 from textwrap import dedent
 
 from pants.backend.codegen.targets.java_thrift_library import JavaThriftLibrary
-from pants.backend.core.targets.resources import Resources
 from pants.backend.jvm.targets.java_library import JavaLibrary
 from pants.backend.jvm.targets.jvm_binary import JvmBinary
 from pants.backend.jvm.targets.scala_library import ScalaLibrary
 from pants.backend.jvm.tasks.jar_create import JarCreate, is_jvm_library
 from pants.build_graph.build_file_aliases import BuildFileAliases
+from pants.build_graph.resources import Resources
 from pants.util.contextutil import open_zip
 from pants_test.jvm.jar_task_test_base import JarTaskTestBase
 from pants_test.tasks.task_test_base import ensure_cached
