+++ tests/python/pants_test/backend/python/tasks/BUILD
@@ -80,7 +80,6 @@ python_tests(
     '3rdparty/python/twitter/commons:twitter.common.dirutil',
     '3rdparty/python:mock',
     ':python_task_test_base',
-    'src/python/pants/backend/core/targets:common',
 
     # TODO(John Sirois): XXX this dep needs to be fixed.  All pants/java utility code needs to live
     # in pants java since non-jvm backends depend on it to run things.
