+++ tests/python/pants_test/core_tasks/test_run_prep_command.py
@@ -9,9 +9,9 @@ import os
 
 from six.moves import range
 
-from pants.backend.core.targets.prep_command import PrepCommand
 from pants.base.exceptions import TaskError
 from pants.build_graph.build_file_aliases import BuildFileAliases
+from pants.build_graph.prep_command import PrepCommand
 from pants.core_tasks.run_prep_command import RunPrepCommand
 from pants.util.contextutil import temporary_dir
 from pants.util.dirutil import touch
